+++ src/index.ts
@@ -7,7 +7,7 @@ import { addProduct, deleteProduct, getProductDetails, getProducts, loadProducts
 import { processNumbers, timeAMPM, fullDateText, truncateText } from "./code/showcase";
 import { totalStakes, stakesCount, transferStake, offerStake, stakesOffered, takeStake, removeStakeOffer } from "./code/stakes";
 import { getTokenData, getTokenRate, tokenOnchainData } from "./code/token";
-import { BlockchainNetwork, ChainIds, EVMAddress, ErrorResponse, MerchantConfigParams, PayTxsData, Payment, PaymentChain, PaymentData, PaymentDataAll, ProductChain, ProductData, ProductDataAll, TxObj } from "./types";
+import { BlockchainNetwork, ChainIds, ChainIdsEnum, EVMAddress, ErrorResponse, MerchantConfigParams, PayTxsData, Payment, PaymentChain, PaymentData, PaymentDataAll, ProductChain, ProductData, ProductDataAll, SUPPORTED_CHAINS, TxObj } from "./types";
 
 export {
     // config
@@ -15,7 +15,9 @@ export {
     getConfig,
     config,
     BlockchainNetwork,
+    SUPPORTED_CHAINS,
     ChainIds,
+    ChainIdsEnum,
     MerchantConfigParams,
 
     // contract address