+++ changes.md
@@ -1,6 +1,10 @@
 
 ## MerchantSlate SDK - Change Log
 
+### 14 September 2025
+* `SUPPORTED_CHAINS` exported constant for list of supported chains
+* `ChainIdsEnum` exported for `ChainIds` type
+
 ### 9 August 2025
 * `getTokenRate` now returns one token unit rate, without required `weiAmount`
 * `tokenRateUSD`, `paidPriceUSD`, `paidTotalUSD`, `paidFeeUSD` added to `loadPayments` returned `paymentsData` array